[Oxybot] Initial commit
Initial commit for oxybot discord bot. This bot is built using JDA.

This commit establishes the structure of the bot by introducing
infrastructure to be used by the bot. This takes the form of the custom
Command infrastructure which adds a Listener class that extends the JDA
ListenerAdapter class. This class enables the bot to begin listening for
events. This ListenerAdapter subclass Listener contains a CommandManager
class which is responsible for keeping track of an executing commands
which are implemented on a per-feature basis.

This commit also adds a test command.